requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.5.0
